Extract dragElement
